Add Trie etc, and move Node to the new file
Trie will hold the root of the trie itself, as well as the slab cache of
preallocated Nodes.

Since I'm defining the Trie in the new file, I think it make sense to
have all of the data structures related to the Trie in that file. Hence
the move of Node.