Add NextNode and fix slab and Node allocation
This finishes the pool/slabs mechanism by allocating all of the Nodes
when the slab is allocated, and growing the pool as necessary.

It changes slab to hold a slice of Node pointers instead of Nodes,
because the latter won't work; the members of Node are a slice and a
map, and they need make() to allocate memory for them.