Replace Node.Words, Node.Children with statically allocatable data structures
Prior to this commit, we use a slice for Words and a map for Children . This
means we have to call make() on each Node when allocating a slab.

In the changes here, we replace the Words slice with a pointer to a linked list
struct, and insert new words at the head for O(1) time in AddWords(). The
default nil value means there are no anagrams at that Node.

We replace the Children member with a separate Node pointer member for each
possible[1] letter. This results in a massive increase in required heap memory,
since the vast majority of Nodes will probably have fewer than three children.

An example of this from one dataset:

    Words   Nodes
        0:  410514
        1:  200556
        2:  12030 (1.92%)
        3:  2239 (0.36%)
        4:  658
        5:  225
        6:  62
        7:  33
        8:  8
        9:  5
        10: 2
        11: 1

    Total Nodes: 626,333

By replacing Children in this way, we are significantly changing the complexity
of finding a child Node. We're going from O(1) lookup in the hash map, to
O(lg n) lookup using the switch statements[2]. Since n == 26, it's possible that
it won't be too much of a slowdown, but early data seem to indicate that this it
will be nontrivial. This may not negatively impact overall execution time, since
memory allocation will likely dominate performance, but it is worth considering.

Simple profiling with the 'go-dict.txt' dataset indicates that execution time
is significantly improved with these changes (from ~1.3s to ~0.95s), number of
allocations is significantly improved (from ~3.3 million to ~1.5 million), and
amount of memory allocated is significantly worse (from ~162MB to ~204MB).
(Note: these results should not be considered definitive or scientific.)

[1]: Where "possible" means "lowercase ASCII a-z". Apologies to Unicode, The Go
     Authors, everyone in the world that dares use characters outside of this
	 archaic range and correspondingly geo-, ethno-, and linguo-centric
     worldview.

     I'm just trying to experiment with making things fast. :(

[2]: https://groups.google.com/d/msg/golang-nuts/6ZeGjRxNT2c/gsObHO9XseMJ

     This link is to a message on golang-nuts from Jesper Louis Andersen from
	 2015-02-28. In it he says that for more than four cases, the compiler
     turns a switch into nested if statements, yielding O(lg n) time.