Replace per-letter Node members and switch{} lookup with static array
Using an array removes the need for the getChild and setChild function
calls, and removes the ugly switches. In theory the previous method was
extensible to include more than the [a-z] character set, but doing so
would be impractical and could never support all Unicode characters. So
it makes more sense to just accept that this approach limits input, and
simplify the implementation.