Add words as []byte and add custom sortedLower()
We read the file into a []byte, but then immediately convert it to
[][]string. Then, when we need to sort each word, we split, sort, and join
it.

In this commit, instead of converting, we split the []byte into [][]byte
on newlines (which, given slice semantics, does not allocate new
memory).

For sorting each []byte word, we first make a lowercase copy, and then
sort with a simple insertion sort.

Most of the performance gains in this commit are due to switching from
string to []byte, but the sort does contribute a nontrivial amount. The
cost of these changes is the same as the others in this branch: going
from Unicode-friendly strings and runes to the very narrow and limited
[a-z]. Hopefully these algorithms and techniques can be backported in a
way that will support Unicode.

For now, though, I may have delved too greedily and too deep.